🗃Refactor: Component.js
