+++ packages/dhrod0325/src/view/constant/Constant.js
@@ -1,20 +1,19 @@
 export const API_SERVER = 'http://localhost:3000';
 
-export const NAV_STATE_EXPLORE = 'STATE_EXPLORE';
-export const NAV_STATE_SAVED = 'NAV_STATE_SAVED';
+export const MAX_IMAGE_NUMBER = 123;
+
+export const IMAGE_API_URL = 'https://randomfox.ca/images';
 
 export const NAV_STATE = {
   EXPLORE: 'NAV_STATE_EXPLORE',
   SAVED: 'NAV_STATE_SAVED',
 };
 
-export const KEY_USER_TOKEN = 'USER_TOKEN';
+export const KEY = {
+  USER_TOKEN: 'USER_TOKEN',
+};
 
 export const EVENT_PIN_NAV_SAVE_CLICKED = 'EVENT_PIN_NAV_SAVE_CLICKED';
 export const EVENT_PIN_NAV_EXPLORE_CLICKED = 'EVENT_PIN_NAV_EXPLORE_CLICKED';
 export const EVENT_PROGRESS_SHOW = 'EVENT_PROGRESS_SHOW';
 export const EVENT_PROGRESS_HIDE = 'EVENT_PROGRESS_HIDE';
-
-export const MAX_IMAGE_NUMBER = 123;
-
-export const IMAGE_API_URL = 'https://randomfox.ca/images';
